Split the webhook show page into focused components.
