+++ resources/js/components/webhook/WebhookLogViewer.vue
@@ -1,0 +1,45 @@
+<script setup lang="ts">
+import { IconWebhook } from '@tabler/icons-vue';
+
+import EmptyState from '@/components/EmptyState.vue';
+import WebhookLogDetail from '@/components/webhook/WebhookLogDetail.vue';
+import WebhookLogList from '@/components/webhook/WebhookLogList.vue';
+import { useWebhookLogs } from '@/composables/useWebhookLogs';
+import type { WebhookLog } from '@/types/webhook';
+
+const props = defineProps<{
+    webhookId: string;
+    logs: WebhookLog[];
+}>();
+
+const { liveLogs, selectedLog, newLogIds, selectLog } = useWebhookLogs(
+    () => props.webhookId,
+    () => props.logs,
+);
+</script>
+
+<template>
+    <div
+        v-if="liveLogs.length > 0"
+        class="grid min-h-0 flex-1 grid-cols-1 overflow-hidden rounded-xl border-2 border-foreground bg-card shadow-2xs lg:h-[calc(100vh-24rem)] lg:grid-cols-3"
+    >
+        <WebhookLogList
+            :logs="liveLogs"
+            :selected-id="selectedLog?.id ?? null"
+            :new-log-ids="newLogIds"
+            @select="selectLog"
+        />
+        <WebhookLogDetail
+            v-if="selectedLog"
+            :webhook-id="webhookId"
+            :log="selectedLog"
+        />
+    </div>
+
+    <EmptyState
+        v-else
+        :icon="IconWebhook"
+        :title="$t('webhooks.show.empty_title')"
+        :description="$t('webhooks.show.empty_description')"
+    />
+</template>
